test: make every lib/ source grep live in a static_rules directory
A source grep is useful — it catches what neither the compiler nor a
behaviour test can see: a missing mpv prefetch-playlist line, a whole-object
watch. The problem was never that they exist, it was that they hid inside
files named after a widget. A reader opened mini_player_test.dart expecting
render tests; a refactorer read the red as broken behaviour.

The new rule is placement and naming, not prohibition: a test that reads
lib/ Dart source lives in test/support/ or test/<layer>/static_rules/ and
ends in _static_rule_test.dart. It carries no exception list, so landing it
green meant fixing all 32 offenders: 17 files that were entirely source
greps moved, and the grep assertions in 12 behaviour tests moved out into
five new static_rules files. Nothing was deleted; the suite count is
unchanged apart from the rule's own six tests.

lib/i18n JSON is deliberately not source: the test that checks translation
defaults asserts on product data, not on a frozen refactor.

lyrics_window_empty_state_test moved whole rather than being split — its
second test compares the keys the window consumes (read from source) against
the keys the service pushes (read at runtime), so the two halves share a
harness and cannot be separated without duplicating it.
