+++ AGENTS.md
@@ -108,13 +108,22 @@ Never:
   two halves of the same features.
 - Do not cite `docs/review/execution-log.md` from code, tests or `AGENTS.md`. It
   is history; a cited snapshot becomes an unmaintained live document.
+- Do not hide a `lib/` source grep inside a test named after a widget or a page.
+  A test that reads `lib/` source is a static rule: it belongs in
+  `test/support/` or `test/<layer>/static_rules/`, with a name ending in
+  `*_static_rule_test.dart`. When a behaviour test needs one grep, move the
+  assertion out, not the file.
+  `test/support/static_rule_placement_static_rule_test.dart` enforces this and
+  deliberately carries no exception list.
 - Do not try to "fix" the benign `Failed to update ui::AXTree` Windows log spam —
   it is a known Flutter engine bug (`flutter/flutter#182444`), not an FMP defect.
   See `docs/troubleshooting.md`.
 
-The last five are enforced by tests under `test/support/` and
-`test/ui/static_rules/`, which carry the exception lists. Add a line with a
-reason when you add a legitimate exception; delete it when it goes away.
+Every rule from the image one down to the placement one is enforced by a test
+under `test/support/` or a `static_rules/` directory; those tests carry the
+exception lists. Add a line with a reason when you add a legitimate exception;
+delete it when it goes away. The first three and the last are conventions —
+nothing checks them for you.
 
 ## Architecture
