+++ lib/services/audio/AGENTS.md
@@ -195,8 +195,8 @@ return `true`.
 - The one-second position fallback yields while something is armed, but only for
   a few ticks. It exists because Android loses the completed event in the
   background; standing down for good would trade one bug for another.
-- `audio_backend_static_test.dart` pins the playlist wrapper and the mpv
-  property — dropping either is silent, and the real backends can only be
+- `audio_backend_gapless_static_rule_test.dart` pins the playlist wrapper and
+  the mpv property — dropping either is silent, and the real backends can only be
   verified on device. The follow path going through the shared track-change
   path, which the handoff gate and the starvation watchdog both depend on, is
   covered behaviourally by `audio_controller_next_medium_test.dart`.
