+++ test/data/static_rules/isar_boundary_static_rule_test.dart
