+++ test/services/static_rules/lyrics_window_strings_static_rule_test.dart
