+++ test/services/static_rules/windows_desktop_service_static_rule_test.dart
