+++ test/ui/pages/history/play_history_page_test.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:fmp/data/models/play_history.dart';
 import 'package:fmp/data/models/track.dart';
@@ -86,30 +84,7 @@ void main() {
     });
   });
 
-  group('history page lazy timeline structure', () {
-    test('timeline list does not expand grouped histories with spread map', () {
-      final source = File(
-        'lib/ui/pages/history/play_history_page.dart',
-      ).readAsStringSync();
-      final timelineBody = _methodBody(source, '_buildTimelineList');
-      final dateGroupBody = _methodBody(source, '_buildDateHeader');
-
-      expect(timelineBody, contains('ListView.builder'));
-      expect(dateGroupBody, isNot(contains('...histories.map')));
-      expect(timelineBody, contains('HistoryTimelineRow'));
-    });
-
-    test('timeline rows are keyed by stable date and history ids', () {
-      final source = File(
-        'lib/ui/pages/history/play_history_page.dart',
-      ).readAsStringSync();
-      final timelineBody = _methodBody(source, '_buildTimelineList');
-
-      expect(timelineBody, matches(RegExp(r"ValueKey\(\s*'history-date-")));
-      expect(timelineBody, matches(RegExp(r"ValueKey\(\s*'history-track-")));
-      expect(timelineBody, contains('key:'));
-    });
-
+  group('history timeline rows', () {
     test(
       'buildHistoryTimelineRows keeps date order and skips collapsed tracks',
       () {
@@ -145,23 +120,6 @@ void main() {
       },
     );
   });
-}
-
-String _methodBody(String source, String name) {
-  final match = RegExp(
-    '(?:^|\\n)\\s*[\\w<>?]+\\s+$name'
-    r'\s*\(',
-  ).firstMatch(source);
-  expect(match, isNotNull, reason: 'method $name should exist');
-  final firstBrace = source.indexOf('{', match!.start);
-  var depth = 0;
-  for (var i = firstBrace; i < source.length; i++) {
-    final char = source[i];
-    if (char == '{') depth++;
-    if (char == '}') depth--;
-    if (depth == 0) return source.substring(firstBrace, i + 1);
-  }
-  fail('method $name body did not close');
 }
 
 Track _buildTrack() {
