+++ test/ui/pages/ranking_ui_state_consumption_test.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -97,65 +95,6 @@ void main() {
         );
       },
     );
-
-    test('home rankings select only the initial loading flag', () {
-      final homePageSource = File(
-        'lib/ui/pages/home/home_page.dart',
-      ).readAsStringSync();
-
-      expect(
-        RegExp(
-          r'ref\.watch\(\s*rankingCacheServiceProvider\.select\(\(state\)\s*=>\s*state\.isInitialLoading\)\s*,?\s*\)',
-          dotAll: true,
-        ).hasMatch(homePageSource),
-        isTrue,
-        reason:
-            'Home rankings should not rebuild for unrelated ranking cache state changes.',
-      );
-    });
-
-    test('explore tabs select only source-specific ranking cache fields', () {
-      final source = File(
-        'lib/ui/pages/explore/explore_page.dart',
-      ).readAsStringSync();
-
-      // 只比對 select 片段，不含 provider 名稱：`dart format` 會把長行折在
-      // provider 與 .select 之間，把兩者綁在同一個字串會讓這條測試隨格式化紅燈。
-      expect(
-        source,
-        isNot(contains('ref.watch(rankingCacheServiceProvider);')),
-      );
-      expect(
-        source,
-        contains(
-          'rankingCacheServiceProvider.select((state) => state.isInitialLoading)',
-        ),
-      );
-      expect(
-        source,
-        matches(
-          RegExp(
-            r'\.select\(\s*\(state\) => state\.errorFor\(SourceIds\.bilibili\)',
-          ),
-        ),
-      );
-      expect(
-        source,
-        matches(
-          RegExp(
-            r'\.select\(\s*\(state\) => state\.errorFor\(SourceIds\.youtube\)',
-          ),
-        ),
-      );
-      expect(
-        source,
-        matches(
-          RegExp(
-            r'\.select\(\s*\(state\) => state\.errorFor\(SourceIds\.netease\)',
-          ),
-        ),
-      );
-    });
   });
 }
