+++ test/ui/static_rules/player_page_static_rule_test.dart
